updating dictionary with normScore

+++ component-dictionary/src/main/groovy/clinicalnlp/dict/stringdist/DynamicStringDist.groovy
@@ -1,6 +1,6 @@
 package clinicalnlp.dict.stringdist
 
 interface DynamicStringDist extends StringDist {
-	Double push(final char c)
+	Float push(final char c)
 	void pop()
 }
